requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.127.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.7.0
fastar==0.8.0
filelock==3.20.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
nodeenv==1.10.0
platformdirs==4.5.1
pre-commit==3.8.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
ruff==0.3.7
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.27.1
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
